moved ethernet back into its own folder
