+++ library.properties
@@ -1,10 +1,9 @@
-name=RoveComm2
-version=2.0.0
+name=RoveComm
+version=3.0.0
 author=MST MRDT
 maintainer=Andrew Van Horn <asvdgv@mst.edu>
 sentence=RoveComm
-paragraph=RoveComm UDP packet protocol
+paragraph=RoveComm packet protocol
 category=Device Control
-url=http://energia.nu/reference/libraries/
-architectures=*
-
+url=https://github.com/MissouriMRDT/RoveComm_Arduino.git
+architectures=teensy,tivac
