+++ src/Platform.h
@@ -1,17 +1,17 @@
 #ifndef ROVECOMM_PLATFORM_H
 #define ROVECOMM_PLATFORM_H
 
-#define TEENSY41 // for testing
-//#include "IPAddress.h"
-
 #if defined(ENERGIA)
 #define ROVECOMM_TIVA 1
 
-#elif defined(TEENSY41)
+#elif defined(ARDUINO_TEENSY41)
 #define ROVECOMM_TEENSY 1
 
 #elif defined(ESP32)
-#define ROVECOMM_ESP32
+#define ROVECOMM_ESP32 1
+
+#else
+#error "Warning: RoveComm may not be supported on this platform!"
 
 #endif // Platform
