+++ src/RoveComm.h
@@ -1,8 +1,8 @@
 #ifndef ROVECOMM_H
 #define ROVECOMM_H
 
-#include "RoveCommEthernet.h"
-//#include "RoveCommEthernetUDP.h"
-//#include "RoveCommEthernetTCP.h"
+#include "Platform.h"
+#include "RoveCommManifest.h"
+#include "RoveCommEthernet/RoveCommEthernet.h"
 
 #endif // ROVECOMM_H
